runner/requirements.txt
ultralytics
kafka-python
opencv-python
sqlalchemy